Update FFG-SWRPG-Dice-Roller.js
Made the code a bit prettier. Tried locate the place where the GM stuff was being used and put it into one location so that if I need to switch the GM name or get the GM name again I can just call on the constant and reduce risk of typos or anything like that.