Update FFG-SWRPG-Dice-Roller.js
The logs did not work because the variable names were incorrect. I updated the logic for the logging so that it can actually detect whether its in logging mode or not. I also ran into a bunch of problems with rolling on the character sheet while not having the -DicePool character on my game. That should be fixed too.